Implement ChannelData interface. This will help in writing general code for new channelType addition
