--- notifications/src/main/kotlin/com/amazon/opendistroforelasticsearch/commons/notifications/model/Webhook.kt
@@ -1,105 +1,0 @@
-/*
- * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- *
- */
-package com.amazon.opendistroforelasticsearch.commons.notifications.model
-
-import com.amazon.opendistroforelasticsearch.notifications.util.logger
-import com.amazon.opendistroforelasticsearch.notifications.util.validateUrl
-import org.elasticsearch.common.Strings
-import org.elasticsearch.common.io.stream.StreamInput
-import org.elasticsearch.common.io.stream.StreamOutput
-import org.elasticsearch.common.io.stream.Writeable
-import org.elasticsearch.common.xcontent.ToXContent
-import org.elasticsearch.common.xcontent.XContentBuilder
-import org.elasticsearch.common.xcontent.XContentParser
-import org.elasticsearch.common.xcontent.XContentParserUtils
-import java.io.IOException
-
-/**
- * Data class representing Webhook channel.
- */
-data class Webhook(
-    val url: String
-) : Writeable, ToXContent {
-
-    init {
-        require(!Strings.isNullOrEmpty(url)) { "URL is null or empty" }
-        validateUrl(url)
-    }
-
-    companion object {
-        private val log by logger(Webhook::class.java)
-        private const val URL_TAG = "url"
-
-        /**
-         * reader to create instance of class from writable.
-         */
-        val reader = Writeable.Reader { Webhook(it) }
-
-        /**
-         * Creator used in REST communication.
-         * @param parser XContentParser to deserialize data from.
-         */
-        @JvmStatic
-        @Throws(IOException::class)
-        fun parse(parser: XContentParser): Webhook {
-            var url: String? = null
-
-            XContentParserUtils.ensureExpectedToken(
-                XContentParser.Token.START_OBJECT,
-                parser.currentToken(),
-                parser
-            )
-            while (parser.nextToken() != XContentParser.Token.END_OBJECT) {
-                val fieldName = parser.currentName()
-                parser.nextToken()
-                when (fieldName) {
-                    URL_TAG -> url = parser.text()
-                    else -> {
-                        parser.skipChildren()
-                        log.info("Unexpected field: $fieldName, while parsing Webhook destination")
-                    }
-                }
-            }
-            url ?: throw IllegalArgumentException("$URL_TAG field absent")
-            return Webhook(url)
-        }
-    }
-
-    /**
-     * Constructor used in transport action communication.
-     * @param input StreamInput stream to deserialize data from.
-     */
-    constructor(input: StreamInput) : this(
-        url = input.readString()
-    )
-
-    /**
-     * {@inheritDoc}
-     */
-    override fun writeTo(output: StreamOutput) {
-        output.writeString(url)
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    override fun toXContent(builder: XContentBuilder?, params: ToXContent.Params?): XContentBuilder {
-        builder!!
-        return builder.startObject()
-            .field(URL_TAG, url)
-            .endObject()
-    }
-}
